+++ src/voices.cpp
@@ -30,7 +30,7 @@ void Voice::read_from_file(char * filename, std::vector<StereoSample> * outsampl
     std::string line;
     std::vector<std::string> tokens;
 
-    uint64_t counter = 0;
+    size_t counter = 0;
 
     std::vector<StereoSample> samples = {};
 
@@ -227,7 +227,7 @@ void Voice::read_from_file(char * filename, std::vector<StereoSample> * outsampl
             for (int i = 0; i < durations[0]*SAMPLING_RATE/speed; i++) { // Note how the speed only affects the duration, not effects
                 stereosample.l = 0;stereosample.r = 0;
 
-                uint64_t baked_i = i; // Used mainly for vibrato
+                size_t baked_i = i; // Used mainly for vibrato
                 for (int k = 0; k < effects.size(); k++) {
                     baked_i = effects[k].get_through_i_effect(baked_i, i);
                 }
